Test fixes
There are a number of attempts to use Deci in ints, which won't work in
2.7 or 3.x, we fix that.

Description doesn't exist in symposion.  So that fails our tests pretty
hard.  Switch that to Private Abstract.  It's clear these tests have not
been run in a very long time.  So both failures and especially passes
need to be taken with salt.

+++ registrasion/tests/test_cart.py
@@ -85,7 +85,7 @@ class RegistrationCartTestCase(MixInPatches, TestCase):
             prod = inventory.Product.objects.create(
                 name="Product " + str(i + 1),
                 description="This is a test product.",
-                category=cls.categories[i / 2],  # 2 products per category
+                category=cls.categories[int(i / 2)],  # 2 products per category
                 price=Decimal("10.00"),
                 reservation_duration=cls.RESERVATION,
                 limit_per_user=10,
